Changed the --gen-communicated-structures flag to --heterogeneous
As mentioned in some meetings, the opts testing of
--gen-communicated-structures wasn't running into the same problems
that pvm testing was.  It turns out that the cause is that PVM is the
only communication layer that #defines CHPL_COMM_HETEROGENEOUS and
that this needs to be defined in order to see the problems that can
cause problems for PVM.

For this reason (and because I've never liked the name of it), I
replaced the --gen-communicated-structures flag with a
(developer-only) --heterogeneous flag which is intended to imply that
we're compiling for a heterogeneous target architecture.  When this is
set, it generates the communicated types' structures as before and
also adds CHPL_COMM_HETEROGENEOUS to the chpl__header.h file (which in
turn causes the values in _type_structure.c to be used).

Updated OPT_TESTING_FLAGS to reflect this change so that we'll start
seeing failures independently of the PVM testing.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17184 3a8e244f-b0f2-452b-bcba-4c88e055c3ca