Reducing number of files that need to get rebuilt when compiler changes
As Sung commented the other week, when the compiler gets rebuilt, we
rebuild a surprising number of files in the runtime (the reason being
that they #include the header chpl_rt_utils.h which is created by
compiling the Chapel part of the runtime and therefore gets
regenerated whenever the compiler does).

I did a review of the files that #included this header.  Some didn't
need to #include it, so I removed it from those files (in some cases
replacing with files it #included that they did need).

For the other cases, I created a new header file called
chpl_rt_utils_static.h which contains the same function prototypes
that the compiler generates in chpl_rt_utils.h and had the .c files
#include that header instead.  These prototypes obviously shouldn't
change because if they did, the static callsites in our .c code would
become invalid.  To ensure that the two files stay in synch, I changed
the build of chpl_rt_utils.chpl to #include chpl_rt_utils_static.h as
well so that if the two get out of sync developers will get a compiler
warning (treated as an error) when we build the common-chpl directory.

With this change, a change to the compiler should only result in the
runtime/src/common-chpl directory getting rebuilt.  Please let me know
if you find other cases in which more is getting rebuilt than you'd
expect.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17187 3a8e244f-b0f2-452b-bcba-4c88e055c3ca