Extended the optimization to narrow wide references further and refactored its implementation.  The following cases may now be narrowed:
  Wide references defined by GET_MEMBER and GET_MEMBER_VALUE where the
  base class or reference is narrow.

  Wide strings defined by STRING_COPY where the string being copied is
  narrow.

  Wide references or classes used via the CAST, GET_REF, SYNC_INIT,
  SYNC_LOCK, SYNC_DESTROY, and NEW_PRIV_CLASS primitives.

Added more code to distinguish between the wide reference vs. wide
class types to make the results of this optimization more applicable.

General clean-up.

Added support to codegen PRIM_MOVE of strings where the LHS is a wide
string and the RHS is a non-wide string and to codegen
PRIM_GET_PRIV_CLASS if the LHS is not a wide class.  These cases are
only exercised now that more classes and references are narrowed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17190 3a8e244f-b0f2-452b-bcba-4c88e055c3ca