Extended the optimization to narrow wide references to work in the presence of local blocks.  The key is to narrow wide references used in PRIM_LOCAL_CHECK if possible, and then remove the PRIM_LOCAL_CHECK if the reference is narrowed.
Added special handling of the case where a reference of a wide class
is assigned a local class so that such statements are broken into two
allowing the local class to be widened.  This special handling already
applied to the case where a wide reference of a wide class was
assigned a local class.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17208 3a8e244f-b0f2-452b-bcba-4c88e055c3ca