Fixing warning in PVM comm code
[Whoops, I fixed this earlier today, but the commit failed and I
hadn't noticed until the regressions failed]

This commit fixes a warning in the PVM communication code that started
showing up when we had the nightly script set CHPL_DEVELOPER to true.
A pthread ID needs to be cast safely to an int, so had to cast to
intptr_t as an intermediate step.  (I originally tried rewriting in
terms of intptr_t, but PVM expects an int, so figured we might as well
do the potential downcast immediately rather than putting it off).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17211 3a8e244f-b0f2-452b-bcba-4c88e055c3ca