Fixed a bug in numBits/numBytes for bools
Someone in the past few months pointed out that our numBits/numBytes
param functions were incorrect for bools (returning 1 bit and 0 bytes
respectively).  They are also incorrect in that they were only defined
for default-width bools (which have no well-known size) and not for
fixed-width bools.

This commit implements numBits/numBytes for fixed-width bools and
generates a compiler error message for default-width bools.  Arguably,
we could implement the param function for default-width bools by
looking at the value of CHPL_TARGET_PLATFORM and CHPL_TARGET_COMPILER
and knowing what it was going to do, but the current approach seems
likely to be sufficient and portable (and will at least cause people
to complain if they don't like it).

Added some tests of both kinds.

If it's not obvious, I'm cleaning up my mailbox this afternoon.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17216 3a8e244f-b0f2-452b-bcba-4c88e055c3ca