Implement tuple assignment to strided arrays, printing of default arithmetic arrays with negative strides, and iterating over default arithmetic arrays with negative strides.
The new dsiSerialWrite for default arithmetic arrays uses the range iterators
for each dimension to walk the array instead of starting at dsiLow and
stepping by stride. It prints the array in an order as though all strides
were positive.

Removed futures for arrNegDom and writeNegStrideArray and updated
writeNegStrideArray.good to expect the sign of the stride of the array to be
ignored when printing it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17221 3a8e244f-b0f2-452b-bcba-4c88e055c3ca