Removed tmpTable and tmpTable2 from default associative domains and removed some of the reliance on tmpTable from default associative arrays. In the methods DefaultAssociativeDom.dsiSorted, DefaultAssociativeDom._resize and DefaultAssociativeArr.dsiSorted, use a function-local copy of the table instead of using a class field.
Removed "_createSortedTmpTable" and "_destroySortedTmpTable" since the copy
will be created when the dsiSorted method is called and destroyed when the
method finishes.

Renamed BaseDom._preserveArrayElement by making Element plural. This method
walks over a list of BaseArr and calls _preserveArrayElement (singular) on each
array.

Added a test that nests calls to the sorted iterator on an associative domain
that failed when the tmpTable field was used instead of a local copy.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17223 3a8e244f-b0f2-452b-bcba-4c88e055c3ca