runtime/ - Removed the chpl_stm_error function. Using chpl_error instead for all cases. - Fixed minor bugs in the chplstm.h macro definitions. - Added support for non-standard sizes (for example 12 bytes) in stm write operations.
compiler/:
Removed primitives and codegen support that targeted chpl_stm_error. 

compiler/AST/expr.cpp:
Added SVEC case for PRIM_TX_STORE in codegen().

compiler/AST/build.cpp:
Removed the changes made to what I thought was dropping the 'on' 
clause on the floor. Based on discussions with Brad / Steve, seems 
like dropping the on-clause causes all the tasks to execute in 
locale 0 and does not match the original motivation of this change 
(which was to avoid supporting the stm equivalent of chpl_comm_fork).  

compiler/passes/insertTransactions.cpp:
- Temporarily disabled all primitives related to ALLOC and FREE.
- Changed the behavior of how I/O and sync operations are treated
within atomic blocks from removing those primitives from the AST 
to throwing an USR_FATAL error. 
- Fixed a bug in the function cloning mechanism which caused the 
compiler to ignore the rest of the function as soon as it encountered 
a 'halt' function call.  
- Cleaned up error messages and warnings. 

modules/internal:
- Removed atomicSupport config param in all but two places (both are in 
DefaultAssociative.chpl to disable the use of _resize method). 
- General cleanup. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17228 3a8e244f-b0f2-452b-bcba-4c88e055c3ca