Adding support for param 'this' arguments to methods
[paired with Jonathan Turner to implement this]

Motivation: Jonathan was working on converting the 'length' query on
strings from a standalone function to a method, but was stymied by the
fact that we support a specialized 'length' on param strings, but
don't currently have the ability to define methods whose 'this'
targets are param.

Looking at the compiler, it looked like this would be fairly
straightforward to implement, so we dove into it as a learning
exercise and got it working.

The syntax is to add a 'param' keyword after the 'def' keyword to
indicate that 'this' should be 'param'.  For secondary methods, this
looks reasonably natural:

      def param string.foo...

indicates that the 'string' value should be 'param'.  For primary
methods it's not quite as syntactically clear, but still unambiguous:

      class C {
        def param foo...
        ...
      }

Note that because we don't support class records or classes
(currently), defining 'this' to be param in the primary case is not
useful.  (We toyed with only implementing this for secondary methods
but our first attempt at that resulted in shift/reduce problems that
we weren't sure how to fix, and we figured we'd probably want this
capability in the (long-?) long-term anyway).

We implemented this change by modifying the parser to accept the
optional 'param' keyword in function definitions and by using that tag
to change the intent of the 'this' parameter from INTENT_BLANK to
INTENT_PARAM.  For secondary methods this was simple and could be done
in the parser itself.  For primary methods, since the 'this' argument
is added later, we had to tuck the INTENT_* tag into the FnSymbol at
parse time (using a new field) then look it up later when adding the
'this' argument.

One side effect of this change is that we also permit the 'param'
keyword to be applied to standalone functions (ones with no 'this'
argument) where it essentially has no effect (since there is no
'this').  We could make this into an error case, but weren't sure
where the best place to put the error check was (and the current
behavior doesn't seem incorrect, just potentially confusing).

Added some tests that show that this is working properly:

secMethodParamThis : defines param this methods as secondary methods
  and makes sure they work as expected

primMethodParamThis : defines param this methods as primary methods
  and makes sure they work as expected

funParamThis* : defines standalone param functions and locks in their
  current behavior.  These would need to change if we made this case
  result in an error

Also, added the missing 'param' keyword to Jonathan's
stringLengthMethodDef future and unfuturized it.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17243 3a8e244f-b0f2-452b-bcba-4c88e055c3ca