Change the SeedGenerator enumerated type in the Random module into a record with a method to generate a seed with the current time. This is very similar to how it worked prior to revision 16848.
Several of our codes that use the Random module have code like:
const seed = if useRandomSeed then SeedGenerator.currentTime else 31415;

When SeedGenerator.currentTime was an enum value, it would be coerced to int
in the above statement and seed would always be assigned either 1 or 31415.

Remove a .future showing that the statement above wasn't doing what was
intended.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17245 3a8e244f-b0f2-452b-bcba-4c88e055c3ca