Adding initial support for external types
This commit extends the parser to support external type declarations
and adds support for external types that are not definable/necessary
to define in Chapel.  For example:

   _extern type foo;

creates an external type named 'foo' whose definition the Chapel
compiler can't reason about (but it can permit variables of that
type, pass them to functions that expect variables of that type,
etc.)

It's worth noting that my implementation of this is very different
from that of normal type symbols in that I essentially created a new
primitive type representing this external type rather than creating a
VarSymbol and trying to resolve that VarSymbol later (which is
essentially impossible to do since it's external).  This reflects the
way I tend to think of type definitions in Chapel -- that they should
define a TypeSymbol rather than a VarSymbol that happens to be a type
-- but that change was bigger than I was able to (or needed to) take
on at this time, so I'm letting sleeping dogs lie.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17256 3a8e244f-b0f2-452b-bcba-4c88e055c3ca