Make sure we don't try to heap-allocate external variables
This commit fixes a bug in external variables in which --no-local
compilations resulted in trying to put the external variable on the
heap even though the Chapel compiler doesn't have any control over the
variable or its type.  This commit fixes that by having the
makeHeapAllocations() routine skip over variables that have
FLAG_EXTERN set.  This ought to fix the regression to new test
cint1b.chpl that showed up over the weekend.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17270 3a8e244f-b0f2-452b-bcba-4c88e055c3ca