Filing a bug against returning arrays in self-recursive calls w/ default args
This test came out of a report by Aleks Gershaft, who is a student in
Larry Snyder's PMP 524 course.

From the .future:

This test demonstrates that if a function has:

(a) self-recursive calls,
(b) its multiple callsites rely on different use of default arguments, and
(c) it returns an array

we get an internal error complaining that the number of formals and
actuals don't match.  I suspect that what's happening is that in
putting in additional arguments to handle the returning of arrays (and
freeing of their memory) we're not properly analyzing one or more of
the wrapper functions (or the original function)?  Breaking any of the
above criteria makes the program work, as shown by
recRetArr-works*.chpl, each of which breaks one of the three things.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17273 3a8e244f-b0f2-452b-bcba-4c88e055c3ca