Remove mutexes from the tasking layer interface.
Mutexes were apparently originally intended as an internal device for
the tasking+threading layers to use for synchronization.  Over time
they've found use outside the tasking layer.  This mod undoes that.  It
removes the chpl_mutex_t type and the CHPL_MUTEX_*() functions from the
tasking layer.  Within the tasking and threading layers it replaces all
uses of CHPL_MUTEX_*() with equivalent threadlayer_mutex_*() ones.
Outside the tasking layer, in chplmemtrack.c and the comm layer, it
replaces them with equivalent CHPL_SYNC_*() uses.

For now we use CHPL_SYNC_[UN]LOCK() in place of CHPL_MUTEX_[UN]LOCK().
But the sync lock and unlock functions are really intended as internal
functionality in support of readFE() and writeEF() and their ilk, so in
the future we may find some way to use more natural sync operations, the
equivalent of readFE() and writeEF(), instead of CHPL_SYNC_[UN]LOCK().



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17278 3a8e244f-b0f2-452b-bcba-4c88e055c3ca