Snapshot of GMP test code
This is a snapshot of my GMP test code, a port of the standard
gmp-chudnovsky.c available on the web.  As of today, I got it to the
point where it passes the Chapel compiler and starts running into
problems compiling the generated C (particularly with some functions
that return external types by var.  Hmm...).  One small step...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17279 3a8e244f-b0f2-452b-bcba-4c88e055c3ca