This change moves the usage of a variable to after the definition point.
This change moves the usage of a variable to after the definition point. Upon
moving all variable declarations to the defPoint, several tests fail
depending on where said defPoint is in relation to its usage. (This is one
fix in what is likely to be several changes -- all of which are noops if
the declarations remain at the top of the function.) In this case, the change
impacts local variables allocated via the heap; the new call expression, the
usage, should fall after the defPoint.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17281 3a8e244f-b0f2-452b-bcba-4c88e055c3ca