Fix bug resulting from r17278 commit.
Among other things, the r17278 commit replaced CHPL_MUTEX_INIT() calls
with CHPL_SYNC_INIT_AUX() calls.  The FIFO tasking layer implementation
of a sync object contains a pointer to a threadlayer mutex that is used
to protect the sync object from simultaneous access.  This mutex is
dynamically allocated by the CHPL_SYNC_INIT_AUX() call.  Unfortunately,
a problem arises when memory tracking is turned on.  Memory tracking
uses a sync object (was a mutex, before r17278) to protect its internal
data structures.  When CHPL_SYNC_INIT_AUX() is called to initialize this
sync object, it does the allocation call to create the threadlayer mutex
for the sync object, which in turn causes a call to the memory tracking
code, which tries to lock precisely the sync object the memory tracker
is initializing.  This fails, typically with a SIGSEGV.

This mod makes the sync object struct directly contain the threadlayer
mutex that protects it, rather than containing a pointer to that mutex.
As a result, initializing a sync object only requires initializing the
mutex, not allocating it and initializing it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17283 3a8e244f-b0f2-452b-bcba-4c88e055c3ca