Next draft of edits to support Nanox (Nanos++)
Since the last commit related to Nanos, the Nanos team
has incorporated the Chapel tasking interface into their
build tree to simplify their support of it over time.
This commit reflects that change by removing code from
runtime/*/tasks/nanos/ and putting in a placeholder
#include to a file from the Nanos implementation.

The version of Nanos that I used to make this work was
from their git source control system, and so I'm not
checking it in here until they can get a copy that they're
happy with us capturing on sourceforge.  The good news is
that with this latest version, all release examples passed.

This change also removes some unused RUNTIME* variables
from runtime/etc/Make*, and makes threads==none the default
threading layer for nanos.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17284 3a8e244f-b0f2-452b-bcba-4c88e055c3ca