Fix another bug resulting from the r17278 commit.
This is similar to the r17283 commit, submitted yesterday.  There, the
problem was that a sync object's struct contained a pointer to a mutex
that had to be allocated, and the fix was to avoid the allocation by
putting the mutex directly in the struct.  Here the problem is that the
pthreads threading layer's auxiliary info in sync and single objects
contains pointers to pthread condition variables, which have to be
allocated.  And the fix is the same, namely to avoid the allocation by
putting the condition variables directly in the auxiliary information.

While here, I also renamed the type for these condition variables.  It
was chpl_condvar_t.  The "chpl_" prefix implied that the type was
global, at least within the runtime.  I renamed it threadlayer_condvar_t
to reflect the fact that it's actually only intended for use by the
threading layer.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17287 3a8e244f-b0f2-452b-bcba-4c88e055c3ca