(This is work I seem to have completed before my trips, but wasn't able to commit.)
Refactored the optimization to narrow wide references so that it can
handle cycles.  That is, if wide reference w1 is only wide because
wide reference w2 is wide, and w2 is only wide because w1 is wide,
then we will not narrow these wide references.

Cleaned up how debugging and statistics are handled with respect to
#defines in this pass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17291 3a8e244f-b0f2-452b-bcba-4c88e055c3ca