Replaced 23 primitives with extern function declarations:
  PRIM_GET_ERRNO "get_errno"
  "_config_has_value"
  "_config_get_value"
  "fopen"
  "fclose"
  "fprintf"
  "fflush"
  "get_stdin"
  "get_stdout"
  "get_stderr"
  "get_nullfile"
  "_init_timer"
  "_now_timer"
  "_seconds_timer"
  "_microseconds_timer"
  "_now_year"
  "_now_month"
  "_now_day"
  "_now_dow"
  "_now_time"
  "chpl_printMemTable"
  "chpl_printMemStat"
  "chpl_memoryUsed"

This is an intermediate commit in an effort to reduce the number of
primitives that need to be manipulated by the compiler.  Several of
these primitives required special handling in function resolution as
primitives (but not as extern functions), so this is a good step
forward.

In this change, I also prefixed more runtime-visible names with chpl_.

There remain 217 primitives.  By the way, there are 90 pragmas
including the new one introduced with this commit.

Added a new pragma "insert line file info" that can be applied to
extern functions for which the compiler should insert the appropriate
line number and filename arguments.  I tried to do this by using the
_get_user_line and _get_user_file primitives, but this resulted in the
creation of string copies (for wide strings) and the extra malloc was
a problem with respect to the chpl_memoryUsed function.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17302 3a8e244f-b0f2-452b-bcba-4c88e055c3ca