This change adds a new pass to the compiler which declares vars on block.
This change adds a new pass to the compiler which repositions all the
definition points of each variable to the top of the block in which it is
used. In the case where the variable is used in two or more blocks, it is
repositioned to the point at which it is common to both of them. The
algorithm is fairly well defined in the comments to the new file
repositionDefExpressions.cpp. The pass is off (for now) by default. Also,
a forthcoming submit will add CondStmts (which aren't considered). Right now,
performance is pretty poor, but it passes the regression suite (with a few
timeouts which work without the timeouts).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17310 3a8e244f-b0f2-452b-bcba-4c88e055c3ca