Removed the "size" field from the default arithmetic array class.
This field was mostly unused, the only real use being in the initialize method
where it computed the value for size, set the field, then used the field.
Changed it into a variable local to the initialize method.

The size field was also not always set correctly, for example array slices
and rank changing slices maintained the size of the original array.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17312 3a8e244f-b0f2-452b-bcba-4c88e055c3ca