Adding support for passing star tuples to external functions
Since a star tuple is implemented by the Chapel compiler as a static
C array, I am currently representing external C arrays as star tuples
within Chapel.  For example:

       _extern type vect3 = 3*int(32);

would represent an external C typedef:

       typedef int32_t vect3[3];

This change permits such vectors to be passed to external C routines
by squashing the & that we usually generate when passing such types
to functions (i.e., passing an int32_t* rather than an int32_t**).

It seems that we ought to be able to remove the & for internal
functions as well now that they're implemented using a C array rather
than a struct, yet trying that change did not work (and surprisingly
did not result in C compile failures, but simply out-of-bounds errors
with really weird index values).  I did not have time to look into
these failures today, but hope to take it up again when I feel GMP is
a little better in the bag.

Added a test showing this capability in action.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17317 3a8e244f-b0f2-452b-bcba-4c88e055c3ca