Updated gmp-chudnovsky
With my last commit, gmp-chudnovsky is now one function call away from
passing the C compilation (for some reason I'm either not getting the
function prototype or am passing the wrong types of actual arguments
in -- I need to diagnose further).  Commenting out the one offending
line permits the compilation to complete and gets a seg fault upon
running.

This updates my gmp-chudnovsky to use the star-tuple trick mentioned
in the previous mail to refer to external C array types.  I've also
added -L and -l lines to the .compopts file to support the linking,
and have added the reference version of the code for...um...reference.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17319 3a8e244f-b0f2-452b-bcba-4c88e055c3ca