This change makes several fixes to the new reposition DefExpr pass.
This change makes several fixes to the new reposition DefExpr pass. Most
importantly, it now accounts for conditionals such that declarations occur
within the scope of a cond. It also considers the block statement info
field as a top level expression (rather than part of the block itself --
consider do...while(var) s.t. 'var' should be declared at the same level as
the do...while as opposed to within the block under the do). All collections
of variables are collected via collectSymExprs (as opposed to treating them
as pure symbols, which wasn't broad enough). Fixed a bug in countBlocks to
make sure I don't count an instance of a variable more than once within a
block (break as soon as I see an instance -- otherwise a variable that
appears multiple times in a single block and nowhere else could have said
variable get declared above that scope).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17332 3a8e244f-b0f2-452b-bcba-4c88e055c3ca