Improved support for passing inout star-tuple arguments to extern functions
With this change, if an external function is expecting a static array:

     void foo(int32_t x[3]);

and it plans to modify that array, declaring its prototype using an
inout intent:

     _extern def foo(inout x: 3*int(32));

will cause the compiler to do the right thing.  In particular, the
compiler passes the original array in to the routine rather than
making a copy (with blank intent, it may make a copy if, for example,
the 3*int(32) is an element of an array) and without generating code
that would try to pass an address of an int32_t [3] in.  The trick I
used here was to let functionResolution modify the external function's
formal argument to be a ref as though it was internal in order to make
all subsequent passes handle it correctly, and then to add a special
case in codegen to add an extra pointer dereference in.

Adding a test to lock in this behavior.  Previously this test would
not work if the external intent was not inout (seg faulted and had
valgrind issues) and it would not compile if the intent was inout
(because we generated code passing in an address-of-int[3] rather than
an int[3]).





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17335 3a8e244f-b0f2-452b-bcba-4c88e055c3ca