Removing the .future file and changing the .good file for test_atomic_on_clause.chpl to reflect the recent changes to the compiler.
Removing "local" block from RA update/verfication loop. The local block
converts some of the wide references to narrow references, causing
the insertTransactions pass to ignore them. This is a problem today
because we still don't have a good mechanism in place to differentiate
references to the stack vs references to the heap. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17373 3a8e244f-b0f2-452b-bcba-4c88e055c3ca