Fixed a bug in which the compiler could issue the error
  "control reaches end of function that returns value"

if the function returns a single enumeration value.  This shows up in
internal functions where return statements are not normalized with a
new return symbol (e.g., an internal function created for a
conditional expression).

The check searches for assignments to the returned symbol on all paths
unless the symbol is an immediate or (now) the symbol is an
enumeration constant.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17382 3a8e244f-b0f2-452b-bcba-4c88e055c3ca