Fixed a bug in the generated code wherein we were not dropping the dummy locale argument from the transactional version of wrapon functions.
This bug was not occuring when CHPL_DEVELOPER was turned on but came up
when we tested using the --fast / -O flags. The problem has do with
function pointers and how the lack of C compiler optimizations allowed
passing fewer arguments when invoking a function via a function pointer.

Not 100% sure why we were not seeing warnings when the generated C code
was being compiled. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17391 3a8e244f-b0f2-452b-bcba-4c88e055c3ca