modifications made to testing system via computePerfStats:
when start_test is run, it can expect an optional .graph
when there exists a .perfkeys for a particular file. 

The contents of the .graph should only specify the graph's
intended title, and the label of the y-axis, each on its own line.  

computePerfStats still creates a .dat, but from the .dat and 
.graph also creates a temporary .gpl which gets run through gnuplot.
The resulting .gif by default goes into CHPL_TEST_PERF_DIR
(./perfdat/<machine>)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17427 3a8e244f-b0f2-452b-bcba-4c88e055c3ca