Filing an interesting but low-priority future that J. Claridge stumbed across
From the .future:

This test shows that when we define initialize() as a secondary
method, it has a different behavior than when we define it as a
primary method.

This is really low priority to fix, because we really want to retire
initialize() altogether.  But until we have a better constructor story
for inherited classes, I think it continues to be useful, regrettably.
Rather than spending time fixing this inconsistency, though, I'd
prefer to see the constructor story fixed.

To that end, I'm filing this future as a placeholder reminder should
other people run into this issue before it is retired rather than
something that we should fix.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17456 3a8e244f-b0f2-452b-bcba-4c88e055c3ca