Futures for type and param iterators
This captures a (low-priority) wishlist item I've had for awhile -- for
the language to support type and param iterators.  Seems like it should
not be terribly hard, yet it's also not terribly important.  I filed
these both because I found myself wanting it during some testing yesterday
and because I'm cleaning out my inbox and found an old mail that was
serving as a placeholder for it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17465 3a8e244f-b0f2-452b-bcba-4c88e055c3ca