Fixed a bug in how PRIM_GET_MEMBER_VALUE was being handled in the local reference case. Earlier, the GET_MEMBER_VALUE primitive was generating a TX_LOAD or a TX_LOAD_REF. This turned out to be incorrect for two reasons.
First, generating the TX_LOAD_REF was not the right primitive because the
original expression was doing the following:
    rhs = &(base->member)
The TX_LOAD_REF was loading the value of base->member into the rhs.

Second,	it turned out that the GET_MEMBER_VALUE was also generating
code for structs when the isClass() test in codegen_member returned
false. This meant if the 'base' was the stack, the 'member' also
landed up being on the stack and hence should not be instrumented.
(E.g. in rhs = base.member, &(base.member) is going to be on the stack
frame)

To deal	with this, I added a new test called isBaseOnStack in the STM pass
to determine if	we need	to generate the	transactional version or leave
things as is.

This fix is only partial and still requires the stack references problem
to be addressed. For example, in the context of the current discussion,
codegen member de-references the base pointer if FLAG_REF is set. We need
to still ensure that this address is not on the stack frame. For example,
   CHPL_STM_LOAD(tx, _ret25, ((*local__tmp6)._high), int64_t, ...
*local__tmp6 may be on the stack. 

We no longer generate a load ref for the transactional version of
GET_MEMBER_VALUE primitive.

I still need to ensure the same problem does not arise in 
PRIM_GET_SVEC_MEMBER_VALUE.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17474 3a8e244f-b0f2-452b-bcba-4c88e055c3ca