Fixed a nasty bug in the stm-gtm-comm-gasnet implementation where we were returning from the comm layer sooner than we ought to.
Fixed another nasty bug that was introduced in r17371 when I refactored
the code while implementing non-aligned 4-bit STM read/write operations.

Both these cases were either not happening when running things within
gdb or on the single-locale case.  

Changed how and when we set an unique id for transactions, required
to maintain distributed transactional state. Note, the id is unique only
on a given locale, but is globally unique since we use <localeId, txid>
to identify transactions globally. Transactions that execute on a single
node and perform no remote operations no longer incur this overhead. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17475 3a8e244f-b0f2-452b-bcba-4c88e055c3ca