Update the arithmetic robustness driver to use an enum instead of a string to choose the distribution to use.
Update the nightly script to use the enum instead of a string. Updated two
.prediff scripts to look for the enum.

This was needed because it is difficult to pass around quoted strings in
scripts and keep them properly quoted. If the string loses the quotes, we
now try to parse it. This change seem much better than trying to figure out
the proper set of quotes and backslashes to use in scripts.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17486 3a8e244f-b0f2-452b-bcba-4c88e055c3ca