Adding performance tests for dgemm and other 2-d operations.
- dgemm: as implmented in our current version of hpl
- init: array initialization (with scalar promotion) in various forms
  (e.g, for, forall, whole array, etc.)
- assign: array assignment in various forms (e.g, for, forall, whole
  array, etc.)
- assign_across_locales: assign and init across locales (currently
  only run for 1 locale with --no-local)

The performance versions of init, assign, and assign_across_locales is
run for 2 problem sizes (256x256, and 1024x1024).  The performance
version of dgemm is run for n=m=o=64, and n=m=o=128 (possibly
reasonable values to use in hpl).

For reference, I've also checked in a C version of dgemm.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17489 3a8e244f-b0f2-452b-bcba-4c88e055c3ca