This change adds support for XMT-style loops using an XMT pragma.
This change adds support for XMT-style loops using "#pragma mta for all
streams i of n." The framework consists of two parts: first, to the parser
and build, we add a new primitive, TPRIMITIVELOOP (and a function to use
it, buildPrimitiveLoopStmt). To call it from Chapel code, the programmer
issues __primitive_loop("xmt pragma forall i in n", per_stream_i,
total_streams_n) { <body> } with an integer for per_stream_i and
total_streams_n which the XMT fills in. Note: we could use existing
primitives, such as TFOR, to implement this functionality, though it would
change/add to the meanings of what exists, and possibly lead to some
confusion. Second, to the primitive code and stmt.cpp, we define
PRIM_BLOCK_XMT_PRAGMA_FORALL_I_IN_N, which allows for usage of the
specific primitive loop we're adding for the XMT. Submits following this
will clean up the formatting of the parser (simply aesthetic) and changes
to the module code for DefaultArithmetic to actually use this code.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17501 3a8e244f-b0f2-452b-bcba-4c88e055c3ca