This change instantiates the new XMT pragma into default arithmetic dom/arrs,
This change instantiates the new XMT pragma into default arithmetic domain
and array loops, specific to the XMT. The structure is the same as for non-
XMT systems: in the leader code, create an array of rank elements, defaulting
the size to be the length of that dimension. For the dimension of interest,
the parDim, set that to be what's returned by computeBlock using the XMT
pragma values i and n, the unique identifier and the total number of streams.
This improves XMT performance -- in some cases, substantially.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17517 3a8e244f-b0f2-452b-bcba-4c88e055c3ca