When --count-tokens is used, parse only the files that are specified on the command line. Only the files on the command line are counted and we exit the compiler once the count is printed, so parsing the modules is unnecessary.
When building forall expressions create unresolved symbols for iterator.leader
and iterator.follower instead of using the gLeaderTag and gFollowerTag
variables. This way forall expressions can be parsed before ChapelBase is
parsed.

This fixes the last token count test failure which was caused by parsing the
internal modules twice when they are specified on the command line and then
having multiple config params with the same name.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17545 3a8e244f-b0f2-452b-bcba-4c88e055c3ca