Added support for startStmStats and stopStmStats to start and stop stm stats collection from Chapel programs.
The upside of this approach is that we no longer need to build the runtime. We
can get stm stats at almost no additional cost by using these methods. Further,
maintianing two separate paths, in addition to developer and non-developer paths,
was getting really tedious and confusing. 

The downside of this approach that it is still a bit coarse grained and turns on
stats on all locales. I didn't bother to add the per-locale version of these 
methods because it will affect the way we collect stats for distributed
transactions. This is actually not hard to implement. We need to make the 
chpl_stm_stats flag part of the transaction state that gets passed around when 
performing remote operations. I didn't see much use for this level of fine detail
right now.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@17560 3a8e244f-b0f2-452b-bcba-4c88e055c3ca