Add some futures I wrote while reading the functions spec chapter.
The *Intent*.chpl tests show that passing domains to functions with
in/out/inout is allowed by the compiler but the spec says this is not allowed.

The varFnRetClasses*.chpl tests declare a class and subclass and return an
instance of each on different paths through a var function. These should get
a compiler error because all returned types must match exactly in var
functions. They currently crash at runtime.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@17577 3a8e244f-b0f2-452b-bcba-4c88e055c3ca